pyproject.toml
[tool.poetry]
authors = ["Integrated Data Service - Dissemination <csvcubed@gsscogs.uk>"]
description = "A tool to generate RDF Data Cube style CSV-W cubes from tidy CSV files. Part of the csvcubed family."
exclude = ["docs/*"]
license = "Apache-2.0"
name = "csvcubed"
packages = [
  {include = "csvcubed", from = "src"},
]
readme = "README.md"
version = "0.1.0-dev"

[tool.poetry.dependencies]
pandas = "~=1.3.3"
python = "^3.9"
pyparsing = "==2.4.7"
jsonschema = "^4.4.0"
appdirs = "^1.4.4"
requests = "^2.27.1"
click = ">=8.0.0"
requests-cache = "^0.9.3"
treelib = "^1.6.1"
uritemplate = "^4.1.1"
python-dateutil = "^2.8.2"
csvcubed-pydantic = ">=1.9.0"
csvcubed-models = "0.1.2"

[tool.poetry.dev-dependencies]
csvcubed-devtools = "0.1.1"
requests-mock = "~1.9.3"
pytest-vcr = "~1.0.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pyright]
exclude = [
  ".venv",
  "**/tests",
  "**/.tox",
]
pythonVersion = "3.9"

[tool.poetry.scripts]
csvcubed = 'csvcubed.cli.entrypoint:entry_point'